fixing expandable sections
Looks like if the expandable sections are viewed from the blog site they aren't formatting correctly. I'm not sure why this is, so I'm going to try to add a space between the headers and the info. If it doesn't work I may have to resign to leaving it as-is until I can fix it later.